refactor: move ledger specific methods to ledger.js
