+++ examples/bcoin-getTrustedInput.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const TX = require('bcoin/lib/primitives/tx');
 const bledger = require('../lib/bledger');
 const {LedgerBcoin} = bledger;
 const {Device} = bledger.HID;
@@ -18,7 +19,7 @@ const {Device} = bledger.HID;
     device: device
   });
 
-  const tx = Buffer.from(
+  const tx = TX.fromRaw(Buffer.from(
       '010000000244faa5b7034adc747a055f423326c67942ac980569c20834b39b709'
     + '9d22b797b000000006a473044022069a8603103073a5adb087cbb4a6feef2af719b9d7d6'
     + '7f3053ee761c910b33705022045ef3521ee78232f25e0f0d3ecc8e741fec055111daa7ba'
@@ -29,9 +30,9 @@ const {Device} = bledger.HID;
     + 'f6f5b2472960ea6802e002e012102cce90d5a8a0a37c2df09b5fea38d546b0ba58ce7099'
     + 'ea8127721bea333da6a2affffffff02a0763577000000001976a9147116d0b9f272ee4b6'
     + 'aac688dd99f44bdb3e9539388ac005ed0b2000000001976a9143c9269ba6d1fbf7505d89'
-    + '94ac71b93e7283f2e9188ac00000000', 'hex');
+    + '94ac71b93e7283f2e9188ac00000000', 'hex'));
 
-  const res = await ledgerBcoin.getTrustedInput(tx, 1);
+  const res = await ledgerBcoin.ledger.getTrustedInput(tx, 1);
 
   console.log('Trusted input:', res.toString('hex'));
