+++ lib/bcoin.js
@@ -8,28 +8,22 @@
 
 const assert = require('assert');
 const util = require('./utils/util');
-const utilTX = require('./utils/transaction');
 
 const {Lock} = require('bmutex');
+
 const Network = require('bcoin/lib/protocol/network');
-
-const TX = require('bcoin/lib/primitives/tx');
 const MTX = require('bcoin/lib/primitives/mtx');
-
 const HDPublicKey = require('bcoin/lib/hd/public');
 const Script = require('bcoin/lib/script').Script;
-
 const secp256k1 = require('bcrypto').secp256k1;
-const bufio = require('bufio');
-const {encoding} = bufio;
 
 const LedgerProtocol = require('./protocol');
 const {APDUCommand, APDUResponse} = LedgerProtocol;
 const {Device} = require('./devices/device');
 
-const DEFAULT_PATH = 'm/0\'/0\'/0\'';
+const LedgerBTC = require('./ledger');
 
-const NULL_SCRIPT = new Script();
+const DEFAULT_PATH = 'm/0\'/0\'/0\'';
 
 /**
  * Ledger BTC App with bcoin primitives
@@ -46,6 +40,7 @@ class LedgerBcoin {
 
   constructor(options) {
     this.device = null;
+    this.ledger = null;
     this.path = DEFAULT_PATH;
     this.network = Network.primary;
 
@@ -73,6 +68,8 @@ class LedgerBcoin {
       this.device.set({
         scrambleKey: 'BTC'
       });
+
+      this.ledger = new LedgerBTC(options.device);
     }
 
     if (options.path != null) {
@@ -105,293 +102,18 @@ class LedgerBcoin {
     assert(Array.isArray(path), 'Path must be string or array');
 
     const indexes = path;
-    const command = APDUCommand.getPublicKey(indexes);
-    const responseBuffer = await this.device.exchange(command.toRaw());
-    const response = APDUResponse.getPublicKey(responseBuffer);
-
-    const rawPubkey = response.data.publicKey;
+    const data = await this.ledger.getPublicKey(path);
+    const rawPubkey = data.publicKey;
     const compressedPubkey = secp256k1.publicKeyConvert(rawPubkey, true);
 
     return new HDPublicKey({
       depth: indexes.length,
       childIndex: indexes[indexes.length - 1],
       parentFingerPrint: 0,
-      chainCode: response.data.chainCode,
+      chainCode: data.chainCode,
       publicKey: compressedPubkey,
       network: this.network
     });
-  }
-
-  /**
-   * Get trusted input
-   * @param {bcoin.TX|Buffer} tx
-   * @param {Number} inputIndex
-   * @returns {Buffer} trustedInput
-   * @throws {LedgerError}
-   */
-
-  async getTrustedInput(tx, inputIndex) {
-    assert(this.device);
-
-    if (Buffer.isBuffer(tx))
-      tx = TX.fromRaw(tx);
-
-    assert(TX.isTX(tx));
-
-    const messages = utilTX.splitTransaction(tx);
-
-    // first packet must contain inputIndex
-    const firstMessage = bufio.write(messages[0].length + 4);
-    firstMessage.writeU32BE(inputIndex);
-    firstMessage.writeBytes(messages[0]);
-    messages[0] = firstMessage.render();
-
-    const last = messages.pop();
-
-    let first = true;
-    for (const message of messages) {
-      const packet = APDUCommand.getTrustedInput(message, first);
-      const res = await this.device.exchange(packet.toRaw());
-
-      // check if throws
-      APDUResponse.getTrustedInput(res);
-      first = false;
-    }
-
-    const packet = APDUCommand.getTrustedInput(last, false);
-    const res = await this.device.exchange(packet.toRaw());
-
-    return APDUResponse.getTrustedInput(res).data;
-  }
-
-  /**
-   * Start composing tx
-   * @async
-   * @param {bcoin.MTX} tx - Mutable transaction
-   * @param {Object} [tis={}] - trusted inputs
-   * @param {Boolean} [isNew=false]
-   * @param {Boolean} [witness=false] - is v1 tx
-   * @throws {LedgerError}
-   */
-
-  async hashTransactionStart(tx, tis = {}, isNew, hasWitness) {
-    assert(this.device);
-    assert(MTX.isMTX(tx));
-
-    const packets = [APDUCommand.hashTransactionStart(
-      utilTX.splitVersionInputs(tx),
-      true,
-      isNew,
-      hasWitness
-    )];
-
-    for (const input of tx.inputs) {
-      const prevoutKey = input.prevout.toKey();
-      let buffer;
-
-      if (tis[prevoutKey]) {
-        // Trusted input
-        const scriptSize = input.script.getSize();
-        const scriptVarintSize = encoding.sizeVarint(scriptSize);
-        buffer = bufio.write(2 + tis[prevoutKey].length + scriptVarintSize);
-
-        buffer.writeU8(0x01);
-        buffer.writeU8(tis[prevoutKey].length);
-        buffer.writeBytes(tis[prevoutKey]);
-        buffer.writeVarint(scriptSize);
-      } else if (hasWitness) {
-        // Prevout + Amount
-        const outpointSize = input.prevout.getSize();
-        const amountSize = 8;
-        const scriptSize = input.script.getSize();
-        const scriptVarintSize = encoding.sizeVarint(scriptSize);
-        const coin = tx.view.getCoinFor(input);
-
-        buffer = bufio.write(1 + amountSize + outpointSize + scriptVarintSize);
-
-        buffer.writeU8(0x02);
-        input.prevout.toWriter(buffer);
-        buffer.writeI64(coin.value);
-        buffer.writeVarint(scriptSize);
-      } else {
-        // Prevout
-        const outpointSize = input.prevout.getSize(); // always 36
-        const scriptSize = input.script.getSize();
-        const scriptVarintSize = encoding.sizeVarint(scriptSize);
-        buffer = bufio.write(1 + outpointSize + scriptVarintSize);
-
-        buffer.writeU8(0x00);
-        input.prevout.toWriter(buffer);
-        buffer.writeVarint(scriptSize);
-      }
-
-      packets.push(APDUCommand.hashTransactionStart(
-        buffer.render(),
-        false,
-        isNew,
-        hasWitness
-      ));
-
-      const scripts = utilTX.splitBuffer(
-        input.script.toRaw(),
-        utilTX.MAX_SCRIPT_BLOCK,
-        true
-      );
-      const last = scripts.pop();
-      const sequence = bufio.write(last.length + 4);
-
-      sequence.writeBytes(last);
-      sequence.writeU32(input.sequence);
-
-      scripts.push(sequence.render());
-
-      for (const script of scripts) {
-        packets.push(APDUCommand.hashTransactionStart(
-          script,
-          false,
-          isNew,
-          hasWitness
-        ));
-      }
-    }
-
-    for (const packet of packets) {
-      const res = await this.device.exchange(packet.toRaw());
-      APDUResponse.hashTransactionStart(res);
-    }
-  }
-
-  /**
-   * Nullify other scripts and start sending inputs
-   * @async
-   * @param {bcoin.MTX} mtx - mutable transaction
-   * @param {String} key - Prevout to key
-   * @param {bcoin.Script} prev - prev script for current input
-   * @param {Object} [tis={}] - trusted inputs map[prevoutKey] = trustedInput
-   * @param {Boolean} [isNew=false] - is new transaction
-   * @param {boolean} [witness=false] - is v1 transaction
-   * @throws {LedgerError}
-   */
-
-  async hashTransactionStartNullify(mtx, key, prev, tis = {}, isNew, witness) {
-    assert(this.device);
-    assert(MTX.isMTX(mtx));
-
-    const newTX = mtx.clone();
-    newTX.view = mtx.view;
-
-    // nullify other input scripts
-    for (const input of newTX.inputs) {
-      const prevoutKey = input.prevout.toKey();
-
-      if (prevoutKey === key)
-        input.script = prev;
-      else
-        input.script = NULL_SCRIPT;
-    }
-
-    await this.hashTransactionStart(newTX, tis, isNew, witness);
-  }
-
-  /**
-   * @async
-   * @param {bcoin.TX} tx
-   * @param {String} key - Prevout to Key(String)
-   * @param {bcoin.Script} prev - prev script for current input
-   * @throws {LedgerError}
-   */
-
-  async hashTransactionStartSegwit(tx, key, prev) {
-    assert(this.device);
-    assert(TX.isTX(tx));
-
-    const newTX = tx.clone();
-    newTX.view = tx.view;
-
-    const inputs = [];
-
-    for (const input of newTX.inputs) {
-      const prevoutKey = input.prevout.toKey();
-
-      if (prevoutKey === key) {
-        input.script = prev;
-        inputs.push(input);
-        break;
-      }
-    }
-
-    newTX.inputs = inputs;
-
-    await this.hashTransactionStart(newTX, {}, false, true);
-  }
-
-  /**
-   * Send and verify outputs
-   * @async
-   * @param {bcoin.TX|Buffer} tx
-   * @returns {Boolean[]}
-   * @throws {LedgerError}
-   */
-
-  async hashOutputFinalize(tx) {
-    assert(this.device);
-
-    if (Buffer.isBuffer(tx))
-      tx = TX.fromRaw(tx);
-
-    assert(TX.isTX(tx));
-
-    let size = encoding.sizeVarint(tx.outputs.length);
-
-    for (const output of tx.outputs)
-      size += output.getSize();
-
-    const outputs = bufio.write(size);
-
-    outputs.writeVarint(tx.outputs.length);
-
-    for (const output of tx.outputs)
-      output.toWriter(outputs);
-
-    const messages = utilTX.splitBuffer(outputs.render(),
-      utilTX.MAX_SCRIPT_BLOCK);
-
-    const lastMessage = messages.pop();
-
-    for (const message of messages) {
-      const packet = APDUCommand.hashOutputFinalize(message, true);
-      const res = await this.device.exchange(packet.toRaw());
-      APDUResponse.hashOutputFinalize(res);
-    }
-
-    const lastPacket = APDUCommand.hashOutputFinalize(lastMessage, false);
-    const res = await this.device.exchange(lastPacket.toRaw());
-    return APDUResponse.hashOutputFinalize(res).data;
-  }
-
-  /**
-   * Sign the processed transaction
-   * @async
-   * @param {String|Numbers[]} path
-   * @param {bcoin.TX|Buffer} tx
-   * @param {bcoin.Script.SighashType} type
-   * @returns {Buffer} signed hash
-   * @throws {LedgerError}
-   */
-
-  async hashSign(tx, path, type) {
-    assert(this.device);
-
-    if (typeof path === 'string')
-      path = util.parsePath(path, true);
-
-    if (Buffer.isBuffer(tx))
-      tx = TX.fromRaw(tx);
-
-    const packet = APDUCommand.hashSign(path, tx.locktime, type);
-    const res = await this.device.exchange(packet.toRaw());
-
-    return APDUResponse.hashSign(res).data;
   }
 
   /**
@@ -437,7 +159,7 @@ class LedgerBcoin {
         continue;
 
       const pokey = li.toKey();
-      const trustedInput = await this.getTrustedInput(li.tx, li.index);
+      const trustedInput = await this.ledger.getTrustedInput(li.tx, li.index);
 
       trustedInputs[pokey] = trustedInput;
     }
@@ -453,8 +175,8 @@ class LedgerBcoin {
 
     if (hasWitness) {
       // load transaction with all inputs for caching hashing data
-      await this.hashTransactionStart(tx, {}, true, true);
-      await this.hashOutputFinalize(tx);
+      await this.ledger.hashTransactionStart(tx, {}, true, true);
+      await this.ledger.hashOutputFinalize(tx);
 
       newtx = false;
     }
@@ -536,7 +258,7 @@ class LedgerBcoin {
     }
 
     if (!witness) {
-      await this.hashTransactionStartNullify(
+      await this.ledger.hashTransactionStartNullify(
         tx,
         inputKey,
         prev,
@@ -545,12 +267,16 @@ class LedgerBcoin {
         witness
       );
 
-      await this.hashOutputFinalize(tx);
+      await this.ledger.hashOutputFinalize(tx);
     } else {
-      await this.hashTransactionStartSegwit(tx, inputKey, prev);
+      await this.ledger.hashTransactionStartSegwit(tx, inputKey, prev);
     }
 
-    const sig = await this.hashSign(tx, ledgerInput.path, ledgerInput.type);
+    const sig = await this.ledger.hashSign(
+      tx,
+      ledgerInput.path,
+      ledgerInput.type
+    );
 
     if (redeem) {
       const stack = vector.toStack();
